Move all files to directories in accordance with their type and purpose
