+++ vet-spring/src/components/admin/lists/VetList.jsx
@@ -1,7 +1,7 @@
-import { usePagination } from "../../context/PaginationContext";
-import { Error } from "../Error";
+import { usePagination } from "../../../context/PaginationContext";
+import { Error } from "../../Error";
 
-const ClientList = () => {
+const VetList = () => {
 
     const {
         getPage,
@@ -59,4 +59,4 @@ const ClientList = () => {
     )
 }
 
-export default ClientList;
+export default VetList;
